+++ coretrace/simulation_data/single_element.hpp
@@ -97,8 +97,8 @@ public:
         this->optics_back = op;
     }*/
 
-    virtual int8_t get_group() const override { return this->group; }
-    virtual void set_group(int8_t group)
+    virtual int32_t get_group() const override { return this->group; }
+    virtual void set_group(int32_t group)
     {
         this->group = group;
         return;
@@ -119,7 +119,7 @@ protected:
     // -2 is an element that does not interact with the trace, ie stage or composite elements. 
     // -1 is an ungrouped element, assigned explicitly or if "group" key is missing from JSON. 
     // 0 and above are grouped elements. 
-    int8_t group;
+    int32_t group;
 };
 
 using single_element_ptr = typename std::shared_ptr<SingleElement>;
