+++ coretrace/simulation_runner/optix_runner/OptixCSP/src/core/soltrace_system.cpp
@@ -951,12 +951,12 @@ int8_t SolTraceSystem::get_group(int32_t element_id)
     if (num_groups > 0) {
         for (size_t i = 0; i < m_groups.size() - 1; ++i) {
             if (element_id >= m_groups[i] && element_id < m_groups[i + 1]) {
-                return static_cast<int8_t>(i);
+                return static_cast<int32_t>(i);
             }
         }
 
         if (element_id >= m_groups[m_groups.size() - 1])
-            return static_cast<int8_t>(m_groups.size() - 1);
+            return static_cast<int32_t>(m_groups.size() - 1);
     }
     
     return -1;
